[ABICLOUDPREMIUM-3159]. Fixed.
In the process that retrieves the IPs I assumed the
virtualmachine has always a virtual appliance.

That's not true in imported virtual machines and hence,
that raised a NullPointerException.

Now it returns an empty list of Ips when the virtual machine
is imported, since we don't handle the NICs in imported
virtual machines.